# docs/openapi/baseline.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Sense2Act Backend API",
    "description" : "招投标商机发现后端。契约文档:docs/api-design.md;内部接口(§9)经 X-Internal-Key 认证,不在本 swagger 暴露。",
    "version" : "v0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/admin/investigation-policies" : {
      "get" : {
        "tags" : [ "admin-policy-controller" ],
        "operationId" : "policies",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePolicyView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "admin-policy-controller" ],
        "operationId" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PolicyPut"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePolicyView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/signals/{id}/investigate" : {
      "post" : {
        "tags" : [ "signal-controller" ],
        "operationId" : "investigate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/signals" : {
      "post" : {
        "tags" : [ "internal-signal-controller" ],
        "operationId" : "push",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignalPush"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePushResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/detection-scan-complete" : {
      "post" : {
        "tags" : [ "internal-signal-controller" ],
        "operationId" : "scanComplete",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ScanComplete"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/watch-profiles" : {
      "get" : {
        "tags" : [ "admin-policy-controller" ],
        "operationId" : "listProfiles",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListProfileView"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "admin-policy-controller" ],
        "operationId" : "createProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfilePost"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseProfileView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/signals/{id}/status" : {
      "patch" : {
        "tags" : [ "signal-controller" ],
        "operationId" : "decide",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StatusPatch"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSignalView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/signals" : {
      "get" : {
        "tags" : [ "signal-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "rule_type",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "score_gte",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "org_id",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "date_from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "date_to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "page_size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResultSignalView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/{id}" : {
      "get" : {
        "tags" : [ "signal-controller" ],
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSignalView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/detection-queue" : {
      "get" : {
        "tags" : [ "internal-signal-controller" ],
        "operationId" : "queue_1",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/watch-profiles/{id}" : {
      "delete" : {
        "tags" : [ "admin-policy-controller" ],
        "operationId" : "deleteProfile",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "PolicyPut" : {
        "type" : "object",
        "properties" : {
          "auto_investigate_threshold" : {
            "type" : "number"
          },
          "max_concurrent_investigations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "default_max_rounds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "default_token_budget" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponsePolicyView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PolicyView"
          }
        }
      },
      "PolicyView" : {
        "type" : "object",
        "properties" : {
          "auto_investigate_threshold" : {
            "type" : "number"
          },
          "max_concurrent_investigations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "default_max_rounds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "default_token_budget" : {
            "type" : "integer",
            "format" : "int32"
          },
          "updated_at" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponseMapStringObject" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        }
      },
      "SignalPush" : {
        "type" : "object",
        "properties" : {
          "document" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "hits" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          },
          "score" : {
            "type" : "number"
          },
          "detector" : {
            "type" : "string"
          }
        }
      },
      "ApiResponsePushResult" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PushResult"
          }
        }
      },
      "PushResult" : {
        "type" : "object",
        "properties" : {
          "signal_id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "auto_investigated" : {
            "type" : "boolean"
          }
        }
      },
...
      "ScanComplete" : {
        "type" : "object",
        "properties" : {
          "document_ids" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "detector" : {
            "type" : "string"
          }
        }
      },
...
      "ProfilePost" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseProfileView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ProfileView"
          }
        }
      },
      "ProfileView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "note" : {
            "type" : "string"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "StatusPatch" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseSignalView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/SignalView"
          }
        }
      },
      "DocRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "publish_date" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "OrgRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "SignalView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "document" : {
            "$ref" : "#/components/schemas/DocRef"
          },
          "org" : {
            "$ref" : "#/components/schemas/OrgRef"
          },
          "hits" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          },
          "score" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string"
          },
          "investigation_id" : {
            "type" : "string"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponsePageResultSignalView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResultSignalView"
          }
        }
      },
      "PageResultSignalView" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SignalView"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page_size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ApiResponseListProfileView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProfileView"
            }
          }
        }
      },
...
      "ApiResponseObject" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          }
        }
      }
    }
  }
}